Update the block API hashing algorithm
The changes allow an IV to be set for the hashing algorithm and reuses
the hash of the module name as the IV of the function name. This
effetively changes the operation from `hash(mname) + hash(fname)` to
`hash(mname + fname)`. Reusing the hash of the module name as the IV
instead of adding the result makes the shellcode smaller.